Use direct REST API methods when available vs HTTP requests
When this plugin is used in multisite to share content between
sites, the large majority of requests will be made to sites that
do not require an HTTP connection. Instead, we can take advantage
of the direct REST API methods available to us.

Introduces the `scheme` attribute, which lets a user choose local
or http. We may build this out to help determine HTTP or HTTPS in
the future.