Support explicit "remote" requests
This clarifies the code we were using for the wsuwp_json shortcode
to determine whether or not a local or remote request should be
made. I've moved it to the base class so that it can become
available for other shortcodes as well.

By default, "local" is used for the scheme attribute. If used in a
multisite configuration, we'll always look to see if a local
request can be made. "http" or "https" can be passed to override
this behavior and make an external request at all times. This can
be useful in multisite if a plugin or theme is enabled on that
other site that adds specific functionality to a post that would
not be available with `switch_to_blog()`.

In the process of clarifying this, support for single site is now
provided. We had previously used `get_blog_details()`, which is a
multisite specific function. There is now logic supporting single
site as well.

Fixes #62