Treat only requests to current site as local
Previously, we treated all requests to sites in the multisite
install as "local" rather than making HTTP requests to retrieve
content. Because of how filters are applied, any content filtering
would happen in the context of the current site rather than the
requested site. For the most part this does not cause an issue, but
can create confusing situations when the configurations of sites
start to become less alike.

Moving back to remote requests keeps things simple. :)