Handle posts as objects instead of classes in local results
When a local REST request is made, the data is returned as an
array as opposed to an object like a remote request. This causes
confusion because each use the same filter to allow the modification
of response data.

This converts local results to use an object instead. It will have
an impact on the research.wsu.edu theme, which appears to be the
only one that actually uses the array data at this point.

That theme will be updated at the same time as this change.