Handle `local_` prefixed shortcode attributes
This extends any extended attributes so that they can be applied
to results requested from the local site via `local_count`. As an
example, `[wsuwp_json count=10 local_count=2 local_tag=blue]` will
pull in 2 local posts tagged as "blue" and merge them with the 10
remote results that don't have any tag restrictions.

In refactoring for this, local results are now retrieved with a
local REST request as opposed to a WP Query. This makes it easier
for the data format to match as expected.